+++ src/index.ts
@@ -1,4 +1,3 @@
-import 'semantic-ui-css/semantic.min.css'
 import './SIA.scss'
 
 // export all type definitions